Fixed truncation of radius

+++ Lattice2/Lattice.cpp
@@ -113,7 +113,7 @@ Lattice::Lattice(int setWidth,
     {
         for(int j = -radius; j <=radius; j++)
         {
-            if((i*i + j*j <= radius*radius) && (i*i + j*j > 0))
+            if((double(i)*double(i) + double(j)*double(j) <= radius*radius) && (i*i + j*j > 0))
             {
                 count++;
                 stepI.push_back(i);
@@ -122,6 +122,7 @@ Lattice::Lattice(int setWidth,
         }
     }
     numNeighbors = stepI.size();
+
     
     parasiteOnGrassDeathIncrement = parasiteOnGrassDeathIncrement / numNeighbors;
     forbOnGrassDeathIncrement = forbOnGrassDeathIncrement / numNeighbors;
